Restructuring the app's file system and writing the new navigation scripts
