+++ Project/tinytask/app/(account)/(app)/(tabs)/_layout.tsx
@@ -1,7 +1,7 @@
 import FontAwesome from "@expo/vector-icons/FontAwesome";
 import { Link, Tabs } from "expo-router";
 import { Pressable, useColorScheme } from "react-native";
-import Colors from "../../constants/Colors";
+import Colors from "../../../../constants/Colors";
 
 function TabBarIcon(props: {
   name: React.ComponentProps<typeof FontAwesome>["name"];
